refactor: restructure search components by introducing SearchSection and updating imports
