+++ src/app/ui/sections/index.ts
@@ -3,7 +3,6 @@ export { default as LateralFilters } from './LateralFilters/LateralFilters';
 export { default as MenuMobile } from './MenuMobile/MenuMobile';
 export { default as MobileLateralFilters } from './MobileLateralFilters/MobileLateralFilters';
 export { default as MobileTopFilters } from './MobileTopFilters/MobileTopFilters';
-export { default as TopFilter } from '@/app/ui/components/SearchBySize/SearchBySize';
 export { default as TireCard } from './TireCard/TireCard';
 export { default as TireInformation } from './TireInformation/TireInformation';
 export { default as TireFeatures } from './TireFeatures/TireFeatures';
@@ -13,10 +12,10 @@ export { default as Benefits } from './Benefits/Benefits';
 export { default as Footer } from './Footer/Footer';
 export { default as SearchBar } from './SearchBar/SearchBar';
 export { default as Pagination } from './Pagination/Pagination';
-export { default as SearchBySizeContainer } from './SearchBySizeContainer/SearchBySizeContainer';
-export { default as SearchByTextContainer } from './SearchByTextContainer/SearchByTextContainer';
+export { default as SearchBySizeContainer } from '@/app/ui/sections/SearchSection/SearchSection';
 export { default as InfoSlider } from './InfoSlider/InfoSlider';
 export { default as TitleSection } from './TitleSection/TitleSection';
 export { default as ServicesGrid } from './ServicesGrid/ServicesGrid';
 export { default as InfoCardsSection } from './InfoCardsSection/InfoCardsSection';
-export { default as SearchContainer } from '@/app/ui/sections/SearchContainer/SearchContainer';
+export { default as SearchContainer } from './SearchContainer/SearchContainer';
+export { default as SearchSection } from './SearchSection/SearchSection';
